requirements.txt
openai
python-dotenv
langchain
pydantic
pydub
python-telegram-bot
gTTS
google-cloud-translate


librosa
soundfile